navbar

+++ app/controllers/services_controller.rb
@@ -3,63 +3,57 @@ class ServicesController < ApplicationController
   before_action :set_service, only: [:show, :edit, :update, :destroy]
 
   def index
-    @users = policy_scope(User)
     @services = policy_scope(Service)
-    @search = params["search"]
-    if @search.present?
-      @nationality = @search["nationality"]
-      @language = @search["language"]
-      @speciality = @search["speciality"]
-      @price = @search["price"]
-      
-      if !@nationality.empty? && !@language.empty?
-        @users = User.where(nationality: @nationality, language: @language)
-      elsif !@nationality.empty?
-        @users = User.where(nationality: @nationality)
-      elsif !@language.empty?
-        @users = User.where(language: @language)
-      else
-        @users = User.all
-      end
-
-      @services = []
-
-      @users.each do |user|
-        if !@speciality.empty? && !@price.empty?
-          user.services.where(speciality: @speciality, price: @price).each do |service|
-            @services << service
-          end
-        elsif !@speciality.empty?
-          user.services.where(speciality: @speciality).each do |service|
-            @services << service
-          end
-        elsif !@price.empty?
-          user.services.where(price: @price).each do |service|
-            @services << service
-          end
-        else
-          user.services.each do |service|
-            @services << service
-          end
-        end
+    if user_signed_in?
+      if current_user.nutritionist? 
+        @services = current_user.services
       end
     else
+
+      @users = policy_scope(User)
       @services = policy_scope(Service)
-      @markers = @services.geocoded.map do |service|
-        {
-          lat: service.latitude,
-          lng: service.longitude,
-          info_window: render_to_string(partial: "info_window", locals: { service: service }),
-          # image_url: helpers.asset_url(service.user.photo)
-        }
+      @search = params["search"]
+      if @search.present?
+        @nationality = @search["nationality"]
+        @language = @search["language"]
+        @speciality = @search["speciality"]
+        @price = @search["price"]
+        
+        if !@nationality.empty? && !@language.empty?
+          @users = User.where(nationality: @nationality, language: @language)
+        elsif !@nationality.empty?
+          @users = User.where(nationality: @nationality)
+        elsif !@language.empty?
+          @users = User.where(language: @language)
+        else
+          @users = User.all
+        end
+
+        @services = []
+
+        @users.each do |user|
+          if !@speciality.empty? && !@price.empty?
+            user.services.where(speciality: @speciality, price: @price).each do |service|
+              @services << service
+            end
+          elsif !@speciality.empty?
+            user.services.where(speciality: @speciality).each do |service|
+              @services << service
+            end
+          elsif !@price.empty?
+            user.services.where(price: @price).each do |service|
+              @services << service
+            end
+          else
+            user.services.each do |service|
+              @services << service
+            end
+          end
+        end
+      else
+        @services = policy_scope(Service)
       end
     end
-
-    # if current_user.nutritionist?
-    #   @services = current_user.services
-    # else
-    #   @services = Service.all
-    # end
   end
 
   def show
@@ -86,19 +80,22 @@ class ServicesController < ApplicationController
   end
 
   def edit 
+    authorize @service
   end
 
   def update 
+    authorize @service
     if @service.update(service_params)
-    redirect_to services_path
+      redirect_to services_path
     else
       render :edit
     end
   end
 
   def destroy
+    authorize @service
     if @service.destroy
-    redirect_to services_path(@service)
+      redirect_to services_path(@service)
     else
       render :index
     end
@@ -109,9 +106,9 @@ class ServicesController < ApplicationController
     params.require(:service).permit(:price, :speciality)
   end
 
-  # def user_params
-  #   params[:service].require(:users).permit(:full_name, :photo, :role)
-  # end
+  def user_params
+    params[:service].require(:users).permit(:full_name, :role, :nationality, :language)
+  end
 
   def set_service
     @service = Service.find(params[:id])
